Add HDL peak detector (with bugs)
The HDl peak detector is not working correctly. The HDL code should save the index of the peak, and keep testing for the value of the dataIn until no new peaks are found between a "k" sized window.

What the current code does is check if there are no new higher peaks in a "k" sized window, and then determine that that is a peak. It checks for "higher" peaks, not "possible peaks", as the Matlab reference does