Integrate Fifo inside OFDMSeparator.
After replacing the divider for the product, a bug appeared where the payload counter would trigger for only an OFDM symbol. This happened because before, the divider had a "0" as the denominator, and the payloadCount was always a high number. When the divider was replaced, the payloadCount started at "1", and the code assumed that the payload was already read fully, when in fact the header hadn't been processed yet.